feat: added implementation for Seismic Acquisition Survey api in sdms-V4

+++ app/sdms-v4/src/apis/schema/handler.ts
@@ -16,7 +16,7 @@
 
 import { Config, StorageFactory } from '../../cloud';
 import { Request as expRequest, Response as expResponse } from 'express';
-
+import { Context } from '../../shared/context';
 import { Operation } from './operations';
 import { Parser } from './parser';
 import { Response } from '../../shared';
@@ -24,7 +24,7 @@ import { SearchService } from '../../services/search';
 import { StorageCoreService } from '../../services';
 import crypto from 'crypto';
 
-export class DatasetHandler {
+export class SchemaHandler {
     public static async handler(req: expRequest, res: expResponse, op: Operation) {
         const dataPartition = req.headers[Config.DATA_PARTITION_ID] as string;
         try {
@@ -32,14 +32,14 @@ export class DatasetHandler {
                 Response.writeOK(res, await this.register(req, dataPartition));
             } else if (op === Operation.Get) {
                 Response.writeOK(res, await this.get(req, dataPartition));
-            } else if (op === Operation.GetVersionedDataset) {
+            } else if (op === Operation.GetVersionedSchema) {
                 Response.writeOK(res, await this.getVersion(req, dataPartition));
-            } else if (op === Operation.DeleteDataset) {
+            } else if (op === Operation.DeleteSchema) {
                 Response.writeOK(res, await this.delete(req, dataPartition));
-            } else if (op === Operation.GetAllVersionIDsOfDataset) {
+            } else if (op === Operation.GetAllVersionIDsOfSchema) {
                 Response.writeOK(res, await this.getAllVersionIDs(req, dataPartition));
-            } else if (op === Operation.ListDatasets) {
-                Response.writeOK(res, await this.listDatasets(req, dataPartition));
+            } else if (op === Operation.ListSchemas) {
+                Response.writeOK(res, await this.listSchemas(req, dataPartition));
             }
         } catch (error) {
             console.log(error);
@@ -48,7 +48,7 @@ export class DatasetHandler {
     }
 
     /**
-     * Register a dataset with storage service
+     * Register a Schema with storage service
      * @param req
      * @param dataPartition
      * @returns list of storage service record identifiers
@@ -56,20 +56,22 @@ export class DatasetHandler {
     private static async register(req: expRequest, dataPartition: string): Promise<string[]> {
         const records = await Parser.register(req);
         const recordIds = await StorageCoreService.insertRecords(req.headers.authorization!, records, dataPartition);
-        for (let ii = 0; ii < records.length; ii++) {
-            const bucketId = this.constructBucketID(recordIds[ii].substring(0, recordIds[ii].lastIndexOf(':')));
-            if (!(await StorageFactory.build(Config.CLOUD_PROVIDER, { dataPartition }).bucketExists(bucketId))) {
-                await StorageFactory.build(Config.CLOUD_PROVIDER, { dataPartition }).createBucket(bucketId);
+        if (Context.schemaGroup.hasBulks) {
+            for (let ii = 0; ii < records.length; ii++) {
+                const bucketId = this.constructBucketID(recordIds[ii].substring(0, recordIds[ii].lastIndexOf(':')));
+                if (!(await StorageFactory.build(Config.CLOUD_PROVIDER, { dataPartition }).bucketExists(bucketId))) {
+                    await StorageFactory.build(Config.CLOUD_PROVIDER, { dataPartition }).createBucket(bucketId);
+                }
             }
         }
         return recordIds;
     }
 
     /**
-     * Get a dataset storage record
+     * Get a Schema storage record
      * @param req
      * @param dataPartition
-     * @returns dataset storage record
+     * @returns Schema storage record
      */
     private static async get(req: expRequest, dataPartition: string) {
         const recordId = Parser.get(req);
@@ -77,10 +79,10 @@ export class DatasetHandler {
     }
 
     /**
-     * Get a version of a dataset storage record
+     * Get a version of a Schema storage record
      * @param req
      * @param dataPartition
-     * @returns dataset storage record
+     * @returns Schema storage record
      */
     private static async getVersion(req: expRequest, dataPartition: string) {
         const [recordId, version] = Parser.getVersion(req);
@@ -88,7 +90,7 @@ export class DatasetHandler {
     }
 
     /**
-     * Delete a dataset storage record and associated cloud storage container
+     * Delete a Schema storage record and associated cloud storage container
      * @param req
      * @param dataPartition
      */
@@ -102,9 +104,10 @@ export class DatasetHandler {
                 throw error;
             }
         }
-
-        const bucketID = this.constructBucketID(inputRecordID);
-        await StorageFactory.build(Config.CLOUD_PROVIDER, { dataPartition }).deleteBucket(bucketID);
+        if (Context.schemaGroup.hasBulks) {
+            const bucketID = this.constructBucketID(inputRecordID);
+            await StorageFactory.build(Config.CLOUD_PROVIDER, { dataPartition }).deleteBucket(bucketID);
+        }
     }
 
     /**
@@ -119,13 +122,13 @@ export class DatasetHandler {
     }
 
     /**
-     * Lists the datasets in a data partition using the search service
+     * Lists the Schemas in a data partition using the search service
      * @param req
      * @param dataPartition
-     * @returns list of dataset storage records that match the dataset kind
+     * @returns list of Schema storage records that match the Schema kind
      */
-    private static async listDatasets(req: expRequest, dataPartition: string) {
-        const options = await Parser.listDatasets(req);
+    private static async listSchemas(req: expRequest, dataPartition: string) {
+        const options = await Parser.listSchemas(req);
         return await SearchService.searchOnKind(
             req.headers.authorization,
             dataPartition,
@@ -136,7 +139,7 @@ export class DatasetHandler {
 
     /**
      * Construct a BucketId from a recordId
-     * @param recordID the dataset storage record Id
+     * @param recordID the Schema storage record Id
      * @returns the bucket id
      */
     private static constructBucketID(recordID: string) {
